chore: Remove duplicated utils from `vane-waterfall`
